Refactoring BackupWatcher
